requirements.txt
Django~=4.1.4
djangorestframework~=3.14.0
mysqlclient==2.1.1